fix: tag_id FK needs INT UNSIGNED to match tags.id (increments), not BIGINT

+++ migrations/2026_09_11_000000_create_feed2forum_feeds_table.php
@@ -6,7 +6,7 @@ return Migration::createTable('feed2forum_feeds', function (Illuminate\Database\
     $table->id();
     $table->string('url');
     $table->string('title');
-    $table->unsignedBigInteger('tag_id')->nullable();
+    $table->unsignedInteger('tag_id')->nullable();
     $table->unsignedInteger('publish_limit')->default(5);
     $table->string('status')->default('active');
     $table->timestamps();
